+++ doc/releasenotes.rst
@@ -15,6 +15,9 @@ Release Notes
   Also add corresponding tests.
 * Add n_worker method to the RedisEvalParallelSampler to enable querying of
   the number of connected workers.
+* Add in-memory database support. Useful, e.g., for benchmarking on slow
+  filesystems or with rather slow network connections.
+
 
 0.8.17
 ------